Initializing the prev_grid when allocating addressed the 'branching with uninitialized values' error foudn by valgrind.
this is what was causing the life experiment to find values it couldn't deal in the support grid, values that shouldn't have been there.

the extra logging i put in was helpful because i printed out both the char and the int (see "agent view" in agent.c or swarm.c) and saw that was garbage.

who and why did this work in the past, older compilers and different implementations I'm guessing.  in addition, i bet that the experiments used the grid differently than lie is.